requirements.txt
pydantic
fastlite
python-telegram-bot[job-queue]
pyyaml
